+++ preliz/distributions/__init__.py
@@ -1,7 +1,7 @@
 from .continuous import *
 from .discrete import *
 
-all_continuous = [Beta, Exponential, Gamma, LogNormal, Normal, Student]
+all_continuous = [Beta, Exponential, Gamma, HalfNormal, LogNormal, Normal, Student]
 all_discrete = [Binomial, NegativeBinomial, Poisson]
